Refactor data loader class
